pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["cmyser"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
pillow = "^9.5.0"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
